refactor: add endless motion and update camera component
The Ortho Width value of the SideViewCameraComponent was
changed from 2048 to 3048, as it was difficult for the player
to see upcoming obstacles in the map and react fast enough
to them. This zooms out the camera a bit more.

Endless motion was added by using the event Tick node, but
this may need to be improved (WIP).